chore: 移除无用文件
