--- config/proxy.ts
@@ -1,16 +1,0 @@
-import type { ProxyOptions } from 'vite'
-
-export function createViteProxy(isOpenProxy: boolean, viteEnv: ImportMetaEnv) {
-  if (!isOpenProxy)
-    return
-
-  const proxy: Record<string, string | ProxyOptions> = {
-    '/api': {
-      target: viteEnv.VITE_APP_API_BASE_URL,
-      changeOrigin: true,
-      rewrite: path => path.replace('/api/', '/'),
-    },
-  }
-
-  return proxy
-}
